// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
    ID: { input: string; output: string };
    String: { input: string; output: string };
    Boolean: { input: boolean; output: boolean };
    Int: { input: number; output: number };
    Float: { input: number; output: number };
    Void: { input: any; output: any };
};
// ...
export type AuthInput = {
    email: Scalars["String"]["input"];
    password: Scalars["String"]["input"];
};
// ...
export type CreateProfileInput = {
    first_name?: InputMaybe<Scalars["String"]["input"]>;
    last_name?: InputMaybe<Scalars["String"]["input"]>;
};
// ...
export type CreateUserInput = {
    auth: AuthInput;
    cvsIds: Array<Scalars["String"]["input"]>;
    departmentId?: InputMaybe<Scalars["ID"]["input"]>;
    positionId?: InputMaybe<Scalars["ID"]["input"]>;
    profile: CreateProfileInput;
    role: UserRole;
};
// ...
export enum UserRole {
    Admin = "Admin",
    Employee = "Employee",
}